# api/openapi.yaml
openapi: 3.0.3
info:
  title: Book Store API
  description: |
    Complete REST API specification for the Book Store eCommerce application.
    
    ## Authentication
    Protected endpoints require a JWT Bearer token obtained from `/api/v1/auth/login` or `/api/v1/auth/register`.
    Include it in the request header: `Authorization: Bearer <token>`
    
    ## Phase 1 Notes
    - Payment processing uses a **mock implementation**. No real transactions are made.
    - Password reset tokens are returned directly in the response (no email service in Phase 1).
    - All prices are in **Indian Rupees (INR / ₹)**.
  version: "1.0"
  contact:
    name: Book Store API Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.bookstore.com/api/v1
    description: Production server
...
security: []
paths:
...
  /books/{bookId}:
    get:
      tags: [Books]
      summary: Get full book details
      operationId: getBookById
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: Book detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BookDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /books/{bookId}/related:
    get:
      tags: [Books]
      summary: Get related books by shared genres
      operationId: getRelatedBooks
      parameters:
        - $ref: '#/components/parameters/BookId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: Related books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────
  # AUTHORS
  # ─────────────────────────────────────────────────
  /authors/{authorId}:
    get:
      tags: [Authors]
      summary: Get author profile with their books
      operationId: getAuthorById
      parameters:
        - $ref: '#/components/parameters/AuthorId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Author profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthorDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /authors/{authorId}/follow:
    post:
      tags: [Authors]
      summary: Follow an author
      operationId: followAuthor
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AuthorId'
      responses:
        '200':
          description: Author followed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          authorId:
                            type: integer
                            format: int64
                          followed:
                            type: boolean
                            example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [Authors]
      summary: Unfollow an author
      operationId: unfollowAuthor
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AuthorId'
      responses:
        '200':
          description: Author unfollowed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          authorId:
                            type: integer
                            format: int64
                          followed:
                            type: boolean
                            example: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ─────────────────────────────────────────────────
  # REVIEWS
  # ─────────────────────────────────────────────────
...
  /cart:
    get:
      tags: [Cart]
      summary: Get the authenticated user's cart
      operationId: getCart
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cart with all items and price summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Cart]
      summary: Clear all items from the cart
      operationId: clearCart
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Cart cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cart/items:
    post:
      tags: [Cart]
      summary: Add a book to the cart
      operationId: addToCart
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '201':
          description: Item added to cart
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CartItemCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /cart/items/{cartItemId}:
    put:
      tags: [Cart]
      summary: Update the quantity of a cart item (0 removes it)
      operationId: updateCartItem
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CartItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Cart item updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CartItemCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Cart]
      summary: Remove a specific item from the cart
      operationId: removeCartItem
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CartItemId'
      responses:
        '204':
          description: Cart item removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────
  # CHECKOUT
  # ─────────────────────────────────────────────────
  /checkout/summary:
    get:
      tags: [Checkout]
      summary: Get checkout summary with optional coupon preview
      operationId: getCheckoutSummary
      security:
        - BearerAuth: []
      parameters:
        - name: couponCode
          in: query
          description: Preview coupon discount before placing order
          schema:
            type: string
      responses:
        '200':
          description: Checkout summary with price breakdown
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CheckoutSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /checkout/validate-coupon:
    post:
      tags: [Checkout]
      summary: Validate a coupon code against the current cart
      operationId: validateCoupon
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponValidateRequest'
      responses:
        '200':
          description: Coupon is valid
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CouponValidateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /checkout/place-order:
    post:
      tags: [Checkout]
      summary: Place the order (creates order record, pending payment)
      operationId: placeOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order placed successfully — proceed to payment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderPlaced'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ─────────────────────────────────────────────────
  # PAYMENTS
  # ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register

  # ── PARAMETERS ────────────────────────────────────────────────────────────
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      description: Unique book identifier
      schema:
        type: integer
        format: int64
        example: 12
    AuthorId:
      name: authorId
      in: path
      required: true
      description: Unique author identifier
      schema:
        type: integer
        format: int64
        example: 3
...
    CartItemId:
      name: cartItemId
      in: path
      required: true
      description: Unique cart item identifier
      schema:
        type: integer
        format: int64
        example: 22
...
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      description: Page size (max 50)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 50
...
  responses:
    BadRequest:
      description: Validation failed or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 400
            error: BAD_REQUEST
            message: Validation failed
            errors:
              - field: email
                message: must be a valid email address
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/auth/register
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 401
            error: UNAUTHORIZED
            message: JWT token is missing or invalid
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/cart
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 404
            error: NOT_FOUND
            message: Book with id 999 not found
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/books/999
    Conflict:
      description: Resource already exists (duplicate)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: 422
            error: UNPROCESSABLE_ENTITY
            message: Coupon SAVE100 has expired
            timestamp: "2025-07-21T10:30:00Z"
            path: /api/v1/checkout/validate-coupon

  # ── SCHEMAS ───────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 400
        error:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Science Fiction
        slug:
          type: string
          example: science-fiction

    # ── Book ────────────────────────────────────────────────────────────
    AuthorSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Daniel Reed
    BookSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12
        title:
          type: string
          example: The Joy of Minimalism
        author:
          $ref: '#/components/schemas/AuthorSummary'
        coverImageUrl:
          type: string
          nullable: true
          example: https://cdn.bookstore.com/covers/joy-minimalism.jpg
        shortDescription:
          type: string
          nullable: true
          example: Declutter your life to uncover peace, clarity, and joy.
        format:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
          example: PAPERBACK
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        price:
          type: number
          format: double
          example: 149.00
        estimatedDeliveryDate:
          type: string
          format: date
          example: "2025-07-28"
        averageRating:
          type: number
          format: double
          nullable: true
          example: 4.5
        copiesSold:
          type: integer
          example: 145
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/BookSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            isbn:
              type: string
              nullable: true
              example: "978-0-000-00000-0"
            language:
              type: string
              example: English
            publishedDate:
              type: string
              format: date
              nullable: true
              example: "2023-05-15"
            reviewCount:
              type: integer
              example: 32
            author:
              $ref: '#/components/schemas/AuthorDetail'
            publisher:
              type: object
              nullable: true
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
                  example: ABC Publishers
                website:
                  type: string
                  nullable: true
    PagedBooks:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/BookSummary'

    # ── Author ────────────────────────────────────────────────────────────
    AuthorDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        name:
          type: string
          example: Daniel Reed
        biography:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        isFollowed:
          type: boolean
          example: false
        books:
          $ref: '#/components/schemas/PagedBooks'
...
    CartItem:
      type: object
      properties:
        cartItemId:
          type: integer
          format: int64
          example: 22
        bookId:
          type: integer
          format: int64
          example: 12
        title:
          type: string
          example: The Joy of Minimalism
        coverImageUrl:
          type: string
          nullable: true
        author:
          $ref: '#/components/schemas/AuthorSummary'
        format:
          type: string
          enum: [PAPERBACK, HARDCOVER, EBOOK]
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        unitPrice:
          type: number
          format: double
          example: 149.00
        quantity:
          type: integer
          minimum: 1
          example: 1
        lineTotal:
          type: number
          format: double
          example: 149.00
        estimatedDeliveryDate:
          type: string
          format: date
    Cart:
      type: object
      properties:
        cartId:
          type: integer
          format: int64
          example: 7
        itemCount:
          type: integer
          example: 2
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: double
          example: 508.00
        updatedAt:
          type: string
          format: date-time
    CartItemRequest:
      type: object
      required: [bookId, quantity]
      properties:
        bookId:
          type: integer
          format: int64
          example: 12
        quantity:
          type: integer
          minimum: 1
          example: 1
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
          description: Set to 0 to remove the item
          example: 2
    CartItemCreated:
      type: object
      properties:
        cartItemId:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        quantity:
          type: integer
        lineTotal:
          type: number
          format: double

    # ── Checkout ────────────────────────────────────────────────────────────
    CheckoutSummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: double
          example: 508.00
        taxAmount:
          type: number
          format: double
          example: 61.00
        deliveryCharge:
          type: number
          format: double
          example: 0.00
        discountAmount:
          type: number
          format: double
          example: 100.00
        totalAmount:
          type: number
          format: double
          example: 469.00
        coupon:
          type: object
          nullable: true
          properties:
            code:
              type: string
              example: SAVE100
            discountType:
              type: string
              enum: [FIXED, PERCENTAGE]
            discountValue:
              type: number
              format: double
    CouponValidateRequest:
      type: object
      required: [couponCode]
      properties:
        couponCode:
          type: string
          maxLength: 50
          example: SAVE100
    CouponValidateResponse:
      type: object
      properties:
        couponCode:
          type: string
        discountType:
          type: string
          enum: [FIXED, PERCENTAGE]
        discountValue:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        valid:
          type: boolean
          example: true
...
    PlaceOrderRequest:
      type: object
      required: [deliveryAddress]
      properties:
        couponCode:
          type: string
          nullable: true
          example: SAVE100
        deliveryAddress:
          $ref: '#/components/schemas/DeliveryAddressRequest'
...
    OrderPlaced:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          example: 7
        orderNumber:
          type: string
          example: BST-20250721-00007
        status:
          type: string
          enum: [PLACED]
        subtotal:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        deliveryCharge:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        estimatedDeliveryDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    # ── Payment ────────────────────────────────────────────────────────────
...